Fixed shooter pid

+++ src/main/java/frc/robot/Util/Gains.java
@@ -5,7 +5,7 @@ public class Gains {
 	public final double kI;
 	public final double kD;
 	public final double kF;
-	public final int kIzone;
+	public final double kIzone;
 	public final double kPeakOutput;
 	/**
 	 * 
@@ -16,7 +16,7 @@ public class Gains {
 	 * @param _kIzone Integral Zone
 	 * @param _kPeakOutput Peak output
 	 */
-	public Gains(double _kP, double _kI, double _kD, double _kF, int _kIzone, double _kPeakOutput){
+	public Gains(double _kP, double _kI, double _kD, double _kF, double _kIzone, double _kPeakOutput){
 		kP = _kP;
 		kI = _kI;
 		kD = _kD;
